pmap improvements for smp
diff originally from art@blahonga.org for i386 (back in 2008!), adapted to amd64
and enhanced by fritsch@genua.de, with minor modifications by yours truly.

short description by fritsch@:

"Instead of mapping the PTEs of an inactive pmap in the APTE range, and then
doing an unconditional remote TLB-flush on all CPUs, we just switch to the other
pmap locally. This only causes a local TLB-flush.  I have also reordered some
stuff so that we will usually send the TLB-shootdown-IPIs first, then do some
local stuff, and then wait for the remote TLB-shootdown to finish."

effect on bare metal should be noticeable, particularly when forking. on kvm,
the difference is brutal: a speedup of nearly 5x.